refactor: move tests to directory
